+++ unpub/lib/src/auth/token_service.dart
@@ -1,5 +1,6 @@
 import 'package:logging/logging.dart';
 
+import '../address.dart';
 import 'auth_config.dart';
 import 'auth_store.dart';
 import 'crypto_box.dart';